test: avoid using old emulator logs in startup
I think this will deflake the storage builds.  Consider this:

- The script to startup the emulator uses the log file to discover the
  port where it is running.
- The script launches the emulator in the background and then
  immediately tries to read its log
- The script fails with "cannot connect", which it is only attempted
  after the port is found.
- The script then prints an empty emulator log, which could not happen
  if the port was found!

But really what is going on is that the shell calls `fork(2)` to start
the emulator, but before it gets to call `exec(2)`, or before it even
redirects `stdout` the parent process reads the file, finds the *old*
ports, then tries to connect. That fails, of course, then it prints the
log file, but that might be empty, or incomplete, or completely
different.

I got paranoid and fixed all the other scripts to start emulators.

Here is a log showing this:

```
2021-07-16T00:00:13Z (+362s)
---------------------------------------------------------
|   Running Storage integration tests (with emulator)   |
---------------------------------------------------------
2021-07-16T00:00:13Z (+362s): Launching Cloud Storage emulator on port 0
2021-07-16T00:00:13Z (+362s): Cannot connect to emulator; aborting test.
2021-07-16T00:00:13Z (+362s): curl connection test result.
curl: (7) Failed to connect to localhost port 39413: Connection refused
Killing emulator server [25990] ... 2021-07-16T00:00:13Z (+362s): ===> ccache stats
```

k

I think (this is based on
some amount of speculation) that when we start the GCS emulator for the
second time the If I am right, when we
start the GCS emulator for a second time