refactor(storage): split hash functions vs. validation (#6996)
For the GCS+gRPC plugin we need to split the hash computation from its
validation. With GCS+gRPC we can send the hashes in the last chunk,
which means we cannot do the validation at the same time (we get the
server-side hashes later). Furthermore, for just validation having all
the hashes in a single string was a neat trick, but now we need to
change their format before uploading. Better to separate the
computation, and to keep the hashes in a data structure that is easier
to operate with.