cleanup(storage): use Disable{MD5*,Crc32c*} consistently
All GCS options are implemented as an `absl::optional<T>`, in the case
of `DisableMD5Hash` and `DisableCrc32cChecksum` this is
`absl::optional<bool>`. The MD5 hashes are disabled by default, this is
achieved by having the default constructor for `DisableMD5Hash`
initialize the option to `true`. CRC32C checksums are enabled by
default, the default constructor initializes the optional to
`absl::nullopt` (this is consistent with all the other options).

After this change we always use these options by doing:

```cc
auto value = request.GetOption<Disable*>().value_or(false);
```

This makes the usage consistent, and only the constructor encodes
what the default is.