feat(common): make GCP_LOG(FATAL) terminate execution (#7058)
`GCP_LOG(FATAL)` now calls `std::abort()` during destruction
of its `Logger`, so control will never return to the caller.
This makes it more closely resemble the behavior of Abseil
logging, which eliminates some confusion between authors and
code readers, and which may also be a feature down the line.

This required some refactoring in a few places because the
control flow is more hidden, and `[[unreachable]]` isn't a
standard attribute yet.

Some, but not all, existing `GCP_LOG(FATAL)` invocations
were immediately followed by calls to `std::exit(1)`. These
were removed. The streaming arguments were also purged of
any trailing newlines and flushes, which `GCP_LOG()` already
handles.

Finally, `storage::AssertOptionSuccess()` was changed to
use `GCP_LOG(FATAL)` instead of `ThrowLogicError()`,
because (as implied in the `TODO`) an assertion failure
is not really something that should be caught.

Fixes #5693.