cleanup(storage): correct return value for WriteCallback (#7066)
In libcurl the callback to receive data is called the "write callback",
I believe because this is the callback where you (the application
developer) write data to the destination file. I always found this
confusing: I am reading data from GCS and use the "write callback" for
it.

In this particular case my confusion manifested as returning
`CURL_READFUNC_PAUSE` from the write callback, when one is supposed to
use `CURL_WRITEFUNC_PAUSE`. Both are macros that expand to the same
value, so the confusion was harmless, but let's fix it anyway.