fix(common): quash "control reaches end of non-void function" error
In #7058 we removed a `std::exit(1)` call after `GCP_LOG(FATAL)`
from `CppTypeToString()` now that the `FATAL` logging does not
return.  However, as mentioned in #7058, this obscures the flow
from the compiler, resulting in a spurious "control reaches end
of non-void function" error.

In this case there is no simple way to rework the function to avoid
that diagnosis, so we quash the error by adding an unreachable
`return`, together with a traditional `/*NOTREACHED*/ comment.

Note: Abseil has an `ABSL_INTERNAL_UNREACHABLE` attribute, which
may be exported in the future to better handle situations like this.