feat(storage): discard handles after an error
We keep CURL* handles in a pool, to reduce the cost of setting up TCP
connections and the SSL handshake. When there is a CURL error we
returned these handles to the pool. This is probably a bad idea,
particularly if the error is caused by timeouts or slow downloads, where
we would want to start over with a new connection.