feat(common): always clog GCP_LOG(FATAL) messages (#7087)
It is desirable that FATAL log messages are never discarded,
especially now that GCP_LOG(FATAL) calls `std::abort` (see #7058).

To that end, ...

- Don't allow FATAL logging to be compile-time disabled.

- Always instantiate a `StdClogBackend` in `DefaultLogBackend()`.

- Give that default `StdClogBackend` a "mininum log severity" from
  the first of the following conditions:
    * If `${GOOGLE_CLOUD_CPP_EXPERIMENTAL_LOG_CONFIG}` is valid,
      use the lowest enabled severity.
    * If `${GOOGLE_CLOUD_CPP_ENABLE_CLOG}` is a valid severity,
      use that.
    * If `${GOOGLE_CLOUD_CPP_ENABLE_CLOG}` is otherwise set,
      use the lowest enabled severity (backwards compatibility).
    * If `${GOOGLE_CLOUD_CPP_ENABLE_CLOG}` is unset, use FATAL.

This means that FATAL messages should always be logged. But note
that `LogSink::DisableStdClog()` and `LogSink::ClearBackends()`
can remove that `StdClogBackend`, so they should be used with care
(undoubtedly only from tests).

Also tidy up some tests so that the disposition of `LogBackend`s
is a little clearer.