feat(storage): control data hotness in benchmark (#7090)
Add an option to control how "hot" is the data in the aggregate
throughput benchmark. Recall that the aggregate throughput benchmark
runs in iterations, during which a full dataset is downloaded. With this
option the same dataset can be downloaded multiple times. This can be
used, for example, to configured the benchmark to read a single file of
1GiB a thousand times in each iteration, which approximates a workload
reading "hot" data. In contrast, configuring the benchmark to read a
dataset of 1024 files of 1GiB each just once approximates an application
reading "cold" data.

Also introduced an option to label the data, so it is easier to load
multiple benchmark results into analysis tools (coLab, Python pandas, R,
etc.) without having to parse the output filenames to annonate the
different runs.