ci(check-api): add IAM and BigQuery to check-api build (#7099)
I also needed to teach the script to skip the api-compatibility check when the "expected" file was missing (but it still gets created for the subsequent runs).

I also changed gzip to use -n, otherwise, the timestamp (and file name) are included in the compressed archive, which results in a different md5sum for every run of this script, even when nothing changes. This results in git thinking that the compressed dump files changed even when they didn't.