chore: add release/changes.sh to help automate release CHANGELOG creation (#7102)
This script generates copy-n-pasteable markdown that can serve as the starting point for a human creating the release notes for our releases. It removes "uninteresting" PRs, and outputs markdown with properly linked section headings and PR numbers. A human will then need to manually tweak the output a bit, but this will helpful be a helpful starting point.