feat(GCS+gRPC): initial support for download timeouts (#7108)
Timeouts for downloads (or any streaming operation) are tricky. A
download can take anywhere from a few milliseconds, to tens of thousands
of seconds. For the REST-based implementation we do not use timeouts for
the full operation, we use a "minimum observed bandwidth" requirement.
There does not seem to a an easy way to implement this in gRPC. While I
work in an implementation setting a timeout (of any kind) is a good
start: at least the programs won't hang forever waiting for stalled
downloads.