feat(storage): make benchmark more robust on stalls
Change the benchmark to use a single queue (LIFO) with the set of
objects to download, and have all the threads service this queue.
Previously we separated the objects in groups and assigned them to
threads. The previous approach wastes threads if one (or a small
fraction) of the downloads gets stalled. I think the new approach better
models a well-designed application.