fix(spanner): fix Client::OverlayQueryOptions() to merge correctly
`Client::OverlayQueryOptions()` was misbehaving on two fronts:
- The environment option for `optimizer_statistics_package` was
  using the value of `${SPANNER_OPTIMIZER_VERSION}` instead of
  `${SPANNER_OPTIMIZER_STATISTICS_PACKAGE}`.
- It was not setting an output value for `request_priority` at all.

At least part of the reason for these errors was that the test:
- assumed the behavior could be changed by modifying the environment
  whereas the implementation only read the environment once, and
- only a single `MockConnection`, with accumulated parameter matches,
  was used across all test cases.

So, fix the implementation, and revamp the test to be more direct
(which eliminates the mocking) and to exercise RequestPriority.

Also add a note to `query_options.h` to update `OverlayQueryOptions()`
whenever new options are added.

Finally, add documentation for `${SPANNER_OPTIMIZER_STATISTICS_PACKAGE}`.