fix(GCS+gRPC): return status on stream closure
The GCS+gRPC version of `ObjectReadSource` was behaving differently from
the REST version. It was returning the last bytes received before the
stream closed, which sets `IsOpen()` to return `false`. The caller
checks `IsOpen` and assumes the last returned status indicates whether
there was an error.

It is possible we want the return type to be `StatusAnd`, but that would
require changing all kinds of APIs that we expose for mocking. Better to
have GCS+gRPC and REST behave consistently, if a bit weirdly.