fix(storage): release handles for oauth2 refresh (#7136)
We were caching the `internal::CurlRequest` object to save some
computation and connection setup time, but these requests are made about
once an hour: the peer often resets the connection before we get to
reuse it. Also it creates the impression that we are leaking sockets. It
is cleaner (I think) to release these as soon as possible.