refactor(storage): isolate header parsing to REST (#7137)
REST uses HTTP headers to return some information. We are mainly
interseted in the the object generation (to resume downloads for the
same object), and the data hashes (for data integrity). Previously the
GCS+gRPC implementation attempted to inject these headers into the
results to keep the changes to the library smaller. This proved messy.

Now we return these values in member variables of the
`internal::ReadSourceResult` struct. Each protocol knows how to parse
its messages and put the right information into these new member
variables. The struct gained a couple of constructors for backwards
compatibility, though in `storage::internal::` it is used for mocking
:disappointed: