refactor(build.sh): make build.sh --docker the default
I did some cleanup and refactoring to `ci/cloudbuild/build.sh` to make
it a little nicer to use in the common case. Changes:

1. `--docker` mode is now **the default build mode**, and may be
specified but does not need to be. So now the following are equivalent
```shell
$ build.sh -t asan-pr  # assumes --docker
$ build.sh -t asan-pr --docker
```

2. `--cloud=project` is **a new flag** to run the build on GCB in the
named project.

3. The `--project=name` flag **has been removed**. If you were using
this flag, use `--cloud=name` instead. I suspect this was rarely used.

4. The build script name may now be specified using the `--build` flag.
Previously this script was specified as a non-flag `$1` to the script.
This was inconsistent.

5. Documentation simplified and clarified. Examples now show common use
cases.

Note: I did not invest much effort to make this change _non-breaking_,
becuase it only affects us and not our users. And the only "breaking"
part of this PR may be for those of us who use the script to trigger
builds on on GCB, which _I think_ is not too common. And for this we
only need to start using the `--cloud=project` flag.