fix(statusor)!: remove status() accessor returning non-const lvalue  (#7150)
This function would allow callers to assign to the contained `Status`
and potentially violate class invariants. For example:

```cc
StatusOr<Foo> sor;  // constructed with kUnknown and no Foo
sor.status() = Status{};  // BUG
auto v = sor.value();  // :-|
```

This PR fixes the above API bug by removing the `status()` accessor that
returns the lvalue to non-const T.

I also remove the `status()` overload that returns the rvalue-ref to
const-T, because I don't think it adds any new functionality to the
class, and a "move" isn't actually happening behind the scenes. Also,
the two access that remain (`const &` and `&&`) are the same two that
`absl::StatusOr` [1] has, so we match there.

[1]: https://github.com/abseil/abseil-cpp/blob/bf31a10b65d945665cecfb9d8807702ae4a7fde1/absl/status/statusor.h#L492-L493