test(storage): deflake integration tests
These integration tests are designed to detect file descriptor leaks in
the REST implementation of the storage library. They do not work
reliably with the GCS+gRPC plugin because gRPC is less predictable in
how it releases resources. It does not matter to us, gRPC has its own
tests for resource leaks, we do not need to duplicate that work.