cleanup(generator): use separator lines consistently
In a recent change to the generator I noticed that we sometimes add
multiple separator lines when only one would do, and don't add one
when one is appropriate.  This remedies that.

While spacing isn't that important, we do submit the generated code
to the repo, so it is nice to be as consistent as possible.  A more
clear-cut generation strategy (like always using leading separators,
except for the initial comment block) might help keep things that
way, but I've just done the minimal tweaks here.  (I tweaked up some
spacing in the generator proper, too.)

Also correct some indentations, although this is even less important
as it is typically fixed during a post `clang-format` phase.