feat(spanner): support for JSON as a column type
`google::cloud::spanner::JSON` is just a strongly-typed string.
That is, a `JSON` value can be constructed from, and converted
to a `std::string`. `JSON` values can also be compared (by string)
for equality, and streamed.

There is no syntax checking of JSON strings in the client library.
All checking is done on the server. The user is expected to build
`JSON` values from well-formatted strings, probably produced by
a separate JSON library. Similarly, `JSON` values should be parsed
by that separate library.