feat(spanner): add factory functions for instance/database/backup
With the move to generated APIs for Spanner admin, the results of
of admin operations now only contain the raw resource strings from
the protobufs. So, it is helpful to be able to build `Instance`,
`Database`, and `Backup` objects directly from those strings.

And to complete the picture, introduce `google::cloud::Project`,
which will be helpful to encapsulate the slew of "projects/"
concatenations we've accumulated.