test(spanner): split the backup/restore test cases into two tests
Split the `BackupTest.*` test cases between two test binaries (for
each of the generated and hand-written versions) so that each binary
can run in parallel, reducing the overall runtime.

This also requires sub-partioning the test instances so that no two
`RestoreDatabase()` calls can run in the same instance at the same
time.