refactor(common): versioned namespace includes major, minor, patch
Fixes: https://github.com/googleapis/google-cloud-cpp/issues/5976

This PR changes our inline namespace to include the major, minor, and
patch numbers. It's been discussed that we should not include the patch
number, and that's OK with me if there's a benefit to doing so.

This PR also adds backward compatibility namespace aliases for both
`v1` and `gcpcxxV1` to avoid breaking users. This does introduce the
`gcpcxxV1` alias in non-generated code where it didn't previously exist
(and the same with `v1` in generated code), but this solution seems to
work fine and seems significantly simpler than any alternative that I
was able to think of.

There's no place to use `@deprecated` on these symbols, and indeed they
don't even show up in Doxygen documentation. The only indication of
deprecation is a mention in the `CHANGELOG.md` file that's included in
this PR.