fix(common): resume sending "v" in "gccl" component of API header
Resume sending the  leading "v" in the value of the "gccl" component
of the `X-Goog-Api-Client` header (see #7383).  The Cloud Spanner
frontend's route lookup service uses the version string in a semver
comparison to determine how the request should be routed, and it
currently expects to see a literal "v".  (For the record, the
minimal version supported by the route lookup service is "v1.17.1".)

Changing the route lookup service to not care about the leading "v"
is non-trivial (it uses a single-value map, indexed by language, to
store the minimum supported version, and the value comes from a flag),
and even then the fix will take some time to roll out, so, in the
meantime, we revert to sending the "v".

Eventually eliminating the "v" is still a goal because it makes the
header conform to the specification, and it also makes C++ behave
like the other languages.