test(spanner): relax an expectation about JSON keys in emulator (#7517)
The Spanner emulator is gaining support for JSON values, which means
we need to start changing some test expectations.  We were somewhat
particular about what errors to expect so that we could discern when
the emulator's behavior started to change.

In this test case, where we use a JSON column as a primary key (which
is always an error), we were expecting a parse error on the word
"JSON", but, for now, we're happy with any kind of error.  We might
still get the "INVALID_ARGUMENT" parsing error, or we could get the
true "INVALID_ARGUMENT" primary-key error, or we might even get
"UNIMPLEMENTED" if the emulator has JSON support dynamically disabled.