feat(common): capture compiler versions at call site
Capture the compiler id, version, features (currently just if exceptions
are enabled), and language version when a `google::cloud::Options` is
constructed. Create a test for this, and modify one of the builds to
verify it does capture a different version when the library is used with
a different C++ version.