feat(GCS+gRPC): endpoint overrides with secure credentials (#7572)
The `GOOGLE_STORAGE_GRPC_ENDPOINT` environment variable can be used to
override the default gRPC endpoint. This is mostly used for tests using
the testbench, and thus defaulted to also switching the credentials to
insecure credentials.  With this change the credentials are kept
unchanged unless you also set `CLOUD_STORAGE_EMULATOR_ENDPOINT`. If that
is set you definitely want to use the emulator and the credentials must
be changed.