feat(generator): support service-qualified names in RPC lists
Allow "<service>.<method>" names in the `omitted_rpcs` and
`gen_async_rpcs` lists (in addition to just "<method>") for
cases where two services defined in the same file share an
RPC name, but we want to treat the RPCs differently.

This isn't anything we need now, but is something I considered
during the recent `omitted_services` change, and it seemed best
to provide the groundwork while I had it in mind.