fix(spanner): avoid use-after move bugs
@devbww and I debugged a use-after move issue that boiled down to these
two Row iterators inspecting a data member (e.g., `row_` and `tup_`) on
subsequent invocations of `op++`, after a caller may have rightfully
moved the object. To avoid this issue, we keep an extra bool in both
iterators to avoid the need to inspect the StatusOr data member again.

Note: This issue didn't show up in our code or builds. It showed up only
when trying to make a different change to `Status`.

Kudos to @devbww for helping to track this down.