fix(spanner): GetSingularRow accepts by value (#7589)
`g::c::spanner::GetSingularRow` extracts and moves the singular row in
the given range, and so it should not take the argument by forwarding
reference, because the caller may pass an lvalue, and it
would be surprising if this function moved a value out of the named
lvalue in the calling function.

Accepting the range by value fixes this by either forcing a copy at the
call site if the caller has an lvalue that they don't want to be
modified, or else the caller can `std::move` the argument, in which case
no copies of the row will be performed -- only moves. I think this is
the best way to make the function unsurprising and allow the caller to
avoid all copies.