feat: enable the Pub/Sub Lite library
This makes the Pub/Sub library usable, though it is still marked as
"experimental".

This grows my little script to create new libraries a little bit. I
think I will remove it soon (the script, `create_scaffold.cmake`), but I
want to unblock Pub/Sub Lite, so here we are.