feat: use cloud.google.com for stable doc links
When linking to the documentation of the different protos use the
reference documentation in `https://cloud.google.com/$product/docs`.
This gives us stable links that do not change as we pin to newever
versions of the APIs.  The downside is that the proto description in
`cloud.google.com` may be newer than the version we used to generate the
code, specially for documents generated with older versions of the
library.