feat(common): add internal-only payload support to Status (#7603)
Related to: #7429
Related to: #1912

This PR adds the ability for non-OK `g::c::Status` objects to hold arbitrary payloads. This does not add any new public API to Status; it is all internal. There's no behavior change in this PR, because nothing is currently setting the payloads. A following PR will start using the payload features.

This payload support will allow us to add and extract the `ErrorInfo` protobuf from Status objects that came from gRPC. We can also do the same for errors from REST services that have an "ErrorInfo" equivalent. We can also include extra stringified data to be logged when we print Status objects, without forcing a protobuf dep on the main `Status` class (which would be bad for gRPC. We would also be able to convert to/from `absl::Status` without losing any data if we choose to do so. All of this will be part of followup PRs.