feat(generator): more stable googleapis proto links (#7606)
We want to minimize the number of changes when we change the commit SHA
for googleapis/googleapis.  With this change we use a Doxygen custom
command in the header files. The CMake file defines this command to
include the embedded SHA. The arguments to the command *should* be more
stable, but will change when proto definitions are moved in their
`.proto` file.