feat(common): add tracing for the LRO polling loop
Get a little visibility into the LRO polling loop in the tracing
log, most notably to see how long we are waiting between polls.

Part of #7562.  The issue looks important, and after failing to
diagnose the problem from the two occurrences we've seen so far,
and failing to reproduce it, the next best thing is to try to
get a little bit more data when it happens again.

And also to that end, add a timestamp to the beginning of each
of the Spanner samples.