refactor(common): Status ignores message if code==OK (#7622)
With this PR, `google::cloud::Status` now ignores the `message` it's constructed with if the `code` is OK. This behavior matches `absl::Status` (which we don't use, but FWIW). We believe the previous behavior where the message was _not_ dropped on OK statuses was a bug, because that can result in "OK" Statuses being not equal. Consider:

```cc
Status ok1{StatusCode::kOk, ""};
Status ok2{StatusCode::kOk, "blah"};
assert(ok1.ok() && ok2.ok());
assert(ok1 != ok2);  // OLD BEHAVIOR -- CONSIDERED A BUG
assert(ok1 == ok2);  // NEW BEHAVIOR -- ALL OKs ARE EQUIVALENT
```
This new behavior differs from `grpc::Status`, which _does_ allow message strings on an OK status. But I imagine this is an oversight due to how it was implemented, and not actually desired behavior.

Requiring that all OK statuses are equivalent enables certain simplifications and optimizations, which we would like to enable.

Note: We (and Abseil) already disallow Status "payloads" on OK statuses.