cleanup(generator): retry traits header is internal (#7623)
Move `google/cloud/$service/retry_traits.h` to
`google/cloud/$service/internal`. It only defines internal symbols, it
makes no sense to make it part of our public API (headers in the public
directories are part of the API).

For libraries that are not GA and/or have not been released with this
header I just moved the file. For libraries that have been released as
GA, I created a backwards compatibility header.

Obviously I changed the generator to use this header at its new
location.