refactor: Status hides error data in Impl
Related to: https://github.com/googleapis/google-cloud-cpp/issues/7429

"OK" statuses are the common case, and they don't have message strings,
payload data, etc. Only non-OK Statuses need this extra data, so paying
the cost of these fields for OK statuses is wasteful.

This PR refactors `google::cloud::Status` to hold an opaque `Impl*` that
holds all the error data for non-OK statuses. OK statuses are
represented by `impl_ == nullptr`, which requires no allocations (again,
this is the common case. This allows us to associate whatever data we
need with non-OK Statuses that we need to without hurting the OK case.

Note: We will be attaching more "ErrorInfo" data with non-OK statuses in
an upcoming PR, so this refactoring will help with that.