refactor: optimize Status moving error data to heap-allocated Impl (#7639)
Related to: https://github.com/googleapis/google-cloud-cpp/issues/7429

Background: "OK" statuses are the common case, and they don't have message strings,
payload data, etc. Only non-OK Statuses need this extra data, so paying
the cost of these fields for OK statuses is wasteful.

This PR refactors `google::cloud::Status` to hold an opaque `Impl*` that
holds all the error data for non-OK statuses. OK statuses are
represented by `impl_ == nullptr`, which requires no allocations (again,
this is the common case). This allows us to associate whatever data we
need with non-OK Statuses without hurting the OK case.

**What's the moved-from state?** We continue to not define the moved-from state for
`Status`. It's a valid state, but unspecified. However, the implementation in this PR
is such that **all moved-from `Status` objects are "OK"**. This matches the default state
when a `Status{}` is constructed. Note that this is a **change in behavior** from the code
previously whereby a moved-from Status would retain its `code()` (and therefore `ok()`) state,
while losing its payload and message. 

Note: We will be attaching more "ErrorInfo" data with non-OK statuses in
an upcoming PR, so this refactoring will help with that.