feat: expose and log ErrorInfo if present
Fixes: https://github.com/googleapis/google-cloud-cpp/issues/7429

This PR adds a `google::cloud::ErrorInfo` class, which mirrors
[ErrorInfo][error-info-proto], and can also work with an equivalent
type in a REST world. We expose programmatic access to this via the
`g::c::Status::error_info()` accessor. We also log this data when
streaming a `g::c::Status` if it's present.

As of today, very few (if any) GCP services actually return this proto,
so I don't expect to see it much yet.

[error-info-proto]: https://github.com/googleapis/googleapis/blob/2e680c6be4b8e48c12f006b945b93edd41be8653/google/rpc/error_details.proto#L112