test(spanner): use the last instance config as the PickConfig() default (#7646)
When no instance configs match our `included_prefix - excluded`
criteria, we return one anyway.  Previously we returned the first
one we saw during the enumeration, but now we return the last.
This only makes a practical difference when running against the
Spanner staging service, where we want to favor `us-central1`
over `cloud-devel-global-config`.

In the longer run we probably want to avoid having any criteria
or default favorites encoded in the source, and instead specify
our preferences using configuration data.  But in the short term
this restores the ability to test using the staging service.