cleanup: use script to update external/googleapis
Use a script to update the list of `.proto` files use by each library,
as well as the list of libraries each library depends on.

The more complicated case was the `google-cloud-cpp::logging_type_protos`
library, its name does not match the Bazel rules where it would generate
`*logging_type_type_protos`. I introduced an `INTERFACE` library for
backwards compatibility.

In the bigquery library we are exposing both the `v1beta1/` and `v1/`
versions of `connection.proto`. I will figure out how to clean that up
in a future PR.