feat(generator): link each method input and output types (#7665)
The protobuf documentation seems to use reference-style markdown links
for protos in the comments, i.e., each reference to `Foo` is formatted
as `[Foo][google.blah.blah.v1.Foo]`.  This formatting is very
distracting if left verbatim, but it can be turned into a link if we
define `google.blah.blah.v1.Foo` as a link reference in Doxygen's
markdown.

With this change the input and output types of each method get such
reference link definitions. The definitions themselves are not visible,
but they make the generated Doxygen output easier to read.

This does not solve the problem when the referenced entity is not an
input or output type of the method, or when it references another
method. But it works in enough cases to make the change worthwhile.