cleanup(bigtable): revert generated admin APIs (#7667)
We want to add per-call Options to the Client APIs (#7666). This would be a breaking change for any GA library. This change will not get done by the next release.

If the generated bigtable admin APIs were included in the next release, they would immediately be GA. This means we would have to support a suboptimal API. There is no hurry to generate the bigtable admin APIs, so we can revert them now, and add them back once the API surface is figured out.

Note that the other new libraries added in this release (tasks, secretmanager, pubsublite) are all marked as experimental. We make no promises for maintaining experimental APIs, so it is acceptable to leave them in the repo, even if the surface might change slightly.