feat(generator): generate quickstart scaffold
Modify the generator to also create a scaffold for the quickstart
program. Then I used the generator to create the scaffold for the
`logging`, `pubsublite`, `secretmanager`, and `tasks` libraries.

The generated code requires some editing, but these are fairly small.

These quickstarts are compiled in the `cmake-install` build, where we
can compile them in parallel. This was very useful, it found several
bugs in the generated CMake files.