feat(generator): add options support to generated clients (#7683)
The generated client constructor now takes a defaulted `Options` argument,
which is merged with the default options for the service, and then stored
as a member.

Similarly, the generated client operations also take a defaulted `Options`
argument, which is then merged with the client options from above, and
installed as the prevailing options for the duration of the operation by
instantiating an `OptionsSpan`.

Any module within the scope of that operation, or its continuations, can
then retrieve those prevailing options.  So, for example, a metadata
decorator could add a header based on a per-operation option, or a retry
policy might modify its behavior based on another option.