feat(storage): capture metadata info in downloads
A lot of useful metadata information is included the the metadata
headers for a GCS download over REST, and as an initial `Object` proto
for gRPC. This information is needed to perform consistent reads from
versioned objects (or detect inconsistent reads from objects without
versioning).  I will need this information as part of the changes for
Arrow too.