feat(spanner): spanner::Client construction from Options (#7706)
Add a new `spanner::Client` constructor using `google::cloud::Options`.
This can then be used to merge with per-operation `Options` in a follow
up PR.

`ClientOptions` construction is still supported for compatibility, but
it is immediately converted to an `Options` value.  `QueryOptions`
is still a first-class citizen, however, because of its use in the
`Connection` interface.