chore(generator): colocate client operation overloads
It is nicer to have the signature-extension operations generated
right next to their request-proto overloads.  This is particularly
true in the client header, which we expect customers to read.

The only true edit is `generator/internal/client_generator.cc`.
The remainder of the PR is updated golden files and regenerated
service interfaces.