fix(spanner): switch order of ListBackupOperations() filter conjuncts
Check that the metadata is of the desired type in the left-hand
side of the `ListBackupOperations()` filter AND expression, before
looking at its value.

This appears to be a recent change in behavior for the production
service, which previously accepted the original ordering, but now
would say: "INVALID_ARGUMENT: Invalid ListBackupOperations request."

Also change the filter subexpressions to be exact matches, rather
than just "contains", so as to make the filtering more precise.

Finally, take this opportunity to log the name of running sample,
which makes it easier to delineate the RPC log.

Fixes #7741.